Finished netcat and port initial test
It was verified how to configure docker-compose.yml to export a port
in all swarm nodes and observed the traffic being distributed.
Containers can be accessible by ping from the host but not from
other docker swarm nodes unless they are on the same network and
are accessed directly from inside a container that the network.
Therefore, it will not be possible to run the base container
to setup the system with --net=host, but it will have to use the
same network as the test containers.